requirements.txt
PyMuPDF==1.23.18
python-docx==1.1.0